% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{crop_and_rotate}
\alias{crop_and_rotate}
\title{Crop and rotate data}
\usage{
crop_and_rotate(data, bbox, angle)
}
\arguments{
\item{angle}{}
}
\value{

}
\description{
Function which is looped
}
