% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{crop_circle}
\alias{crop_circle}
\title{Crop sf as circle}
\usage{
crop_circle(data, original_bbox, angle = pi/6)
}
\arguments{
\item{angle}{}
}
\value{

}
\description{
Crop an sf object to a circular shape (or any shape with
even sides).
}
