% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_functions.R
\name{crop_sf}
\alias{crop_sf}
\title{Crop sf}
\usage{
crop_sf(data, n_vertices = 4, radius = 10, crop_center = c(0, 0))
}
\arguments{
\item{data}{sf geometry data.}

\item{n_vertices}{Number of vertices crop will have.}

\item{radius}{Radius of sf geometry crop.}

\item{crop_center}{Central point of sf crop.}
}
\value{
sf geometry data.
}
\description{
Crop an sf geometry object to a desired 4-sided shape. The shape is
defined by a radius and a number of vertices. The radius is the distance
from the center to the tangent point of each face. The crop center is
the central point of the sf geometry point to crop round.
}
