% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{geometry_to_points}
\alias{geometry_to_points}
\title{Geometry to points}
\usage{
geometry_to_points(data)
}
\arguments{
\item{data}{sf geometry data.}
}
\value{
A data.frame.
}
\description{
Converts sf geometry to a data.frame.
Wrapper of the sf::st_coordinates function.
}
