% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{rotate_data}
\alias{rotate_data}
\title{Rotate data}
\usage{
rotate_data(data, rotate_angle = pi/4)
}
\arguments{
\item{rotate_angle}{}
}
\value{

}
\description{
Rotate sf data by the angle defined. You will likely want to
use a value half that of the shear angle if they are used in
conjunction.
}
