% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{shear_data}
\alias{shear_data}
\title{Shear data}
\usage{
shear_data(data, shear_angle = pi/4)
}
\arguments{
\item{shear_angle}{}
}
\value{

}
\description{
Shear data using the shear angle defined.
}
