% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{transpose_bbox}
\alias{transpose_bbox}
\title{transpose boundary box}
\usage{
transpose_bbox(original_bbox)
}
\arguments{
\item{bbox}{}
}
\value{

}
\description{
Centers the bbox at (0,0) for easy rotation cropping.
}
