% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{transpose_data}
\alias{transpose_data}
\title{Transpose data}
\usage{
transpose_data(data, x_add = 0, y_add = 0)
}
\arguments{
\item{y_add}{}
}
\value{

}
\description{
Transpose an sf object by (x_add, y_add).
}
