% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{create_boundary_shape}
\alias{create_boundary_shape}
\title{Create a shape which matches the boundary.}
\usage{
create_boundary_shape(vertices = 24, radius = 1, transpose = c(0, 0))
}
\arguments{
\item{vertices}{Number of vertices.}

\item{radius}{distance from center to each vertex.}

\item{transpose}{(X, Y) distance to transpose by.}
}
\value{
sf POLYGON
}
\description{
Create an sf Polygon of any regular shape.
}
