% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_functions.R
\name{crop_and_rotate}
\alias{crop_and_rotate}
\title{Crop and rotate data}
\usage{
crop_and_rotate(data, bbox, angle)
}
\arguments{
\item{data}{sf geometry data.}

\item{bbox}{An sf::st_bbox-defined object.}

\item{angle}{Angle to rotate by, in Radians.}
}
\value{
sf geometry data.
}
\description{
Crop sf geometry data using an st::bbox and rotate by angle Radians.
}
