% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{transpose_data}
\alias{transpose_data}
\title{Transpose data}
\usage{
transpose_data(data, x_add = 0, y_add = 0)
}
\arguments{
\item{data}{sf geometry data.}

\item{x_add}{Value to transpose data by, on the x-axis.}

\item{y_add}{Value to transpose data by, on the y-axis.}
}
\value{
sf geometry data.
}
\description{
Transpose an sf object by (x_add, y_add).
}
